+++ src/apps/kwenta/optimism/kwenta.isolated.contract-position-fetcher.ts
@@ -1,0 +1,12 @@
+import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
+
+import { OptimismKwentaPerpContractPositionFetcher } from '../common/kwenta.perp.contract-position-fetcher';
+
+@PositionTemplate()
+export class OptimismKwentaIsolatedContractPositionFetcher extends OptimismKwentaPerpContractPositionFetcher {
+  groupLabel = 'Isolated Margin';
+
+  getAccountAddress(address: string): string {
+    return address;
+  }
+}