+++ src/apps/kyberswap-elastic/contracts/index.ts
@@ -5,6 +5,7 @@ import { ContractFactory } from '~contract/contracts';
 import { Network } from '~types/network.interface';
 
 import { Factory__factory } from './ethers';
+import { KyberswapElasticLm__factory } from './ethers';
 import { Pool__factory } from './ethers';
 import { PositionManager__factory } from './ethers';
 import { TickReader__factory } from './ethers';
@@ -13,13 +14,16 @@ import { TickReader__factory } from './ethers';
 type ContractOpts = { address: string; network: Network };
 
 @Injectable()
-export class KyberSwapElasticContractFactory extends ContractFactory {
+export class KyberswapElasticContractFactory extends ContractFactory {
   constructor(@Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit) {
     super((network: Network) => appToolkit.getNetworkProvider(network));
   }
 
   factory({ address, network }: ContractOpts) {
     return Factory__factory.connect(address, this.appToolkit.getNetworkProvider(network));
+  }
+  kyberswapElasticLm({ address, network }: ContractOpts) {
+    return KyberswapElasticLm__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
   pool({ address, network }: ContractOpts) {
     return Pool__factory.connect(address, this.appToolkit.getNetworkProvider(network));
@@ -33,6 +37,7 @@ export class KyberSwapElasticContractFactory extends ContractFactory {
 }
 
 export type { Factory } from './ethers';
+export type { KyberswapElasticLm } from './ethers';
 export type { Pool } from './ethers';
 export type { PositionManager } from './ethers';
 export type { TickReader } from './ethers';