+++ src/apps/lyra-avalon/lyra-avalon.definition.ts
@@ -9,6 +9,14 @@ export const LYRA_AVALON_DEFINITION = appDefinition({
   description:
     'Lyra is an options trading protocol accessing the scalability of Layer 2 Ethereum to provide a robust, lightning-fast and reliable trading experience.',
   url: 'https://avalon.app.lyra.finance/',
+  tags: [AppTag.OPTIONS],
+  keywords: [],
+  primaryColor: '#fff',
+
+  links: {
+    discord: 'https://discord.com/invite/P49mj6UbmC',
+    github: 'https://github.com/lyra-finance',
+  },
 
   groups: {
     options: {
@@ -29,32 +37,22 @@ export const LYRA_AVALON_DEFINITION = appDefinition({
       label: 'Staking',
     },
 
-    ve: {
-      id: 've',
+    stkLyra: {
+      id: 'stk-lyra',
       type: GroupType.TOKEN,
-      label: 'VotedEscrow',
+      label: 'stkLYRA',
     },
 
-    claimable: {
-      id: 'claimable',
+    stkLyraClaimable: {
+      id: 'stk-lyra-claimable',
       type: GroupType.POSITION,
-      label: 'Rewards',
+      label: 'stkLYRA Rewards',
     },
-  },
-
-  tags: [AppTag.OPTIONS],
-  keywords: [],
-
-  links: {
-    discord: 'https://discord.com/invite/P49mj6UbmC',
-    github: 'https://github.com/lyra-finance',
   },
 
   supportedNetworks: {
     [Network.OPTIMISM_MAINNET]: [AppAction.VIEW],
   },
-
-  primaryColor: '#fff',
 });
 
 @Register.AppDefinition(LYRA_AVALON_DEFINITION.id)
@@ -63,5 +61,3 @@ export class LyraAvalonAppDefinition extends AppDefinition {
     super(LYRA_AVALON_DEFINITION);
   }
 }
-
-export default LYRA_AVALON_DEFINITION;