+++ src/apps/mean-finance/polygon/mean-finance.dca-position.contract-position-fetcher.ts
@@ -6,7 +6,7 @@ import { Register } from '~app-toolkit/decorators';
 import { getImagesFromToken, getLabelFromToken } from '~app-toolkit/helpers/presentation/image.present';
 import { ContractType } from '~position/contract.interface';
 import { PositionFetcher } from '~position/position-fetcher.interface';
-import { ContractPosition } from '~position/position.interface';
+import { ContractPosition, Standard } from '~position/position.interface';
 import { claimable, supplied } from '~position/position.utils';
 import { Network } from '~types/network.interface';
 
@@ -23,6 +23,7 @@ type MeanFinanceDcaPositionContractPositionDataProps = {
   from: string;
   to: string;
   positionId: string;
+  assetStandard: Standard;
 };
 
 @Register.ContractPositionFetcher({ appId, groupId, network })
@@ -67,6 +68,7 @@ export class PolygonMeanFinanceDcaPositionContractPositionFetcher implements Pos
           to: to.address,
           liquidity,
           positionId: dcaPosition.id,
+          assetStandard: Standard.ERC_721,
         },
         displayProps: {
           label,
@@ -97,6 +99,7 @@ export class PolygonMeanFinanceDcaPositionContractPositionFetcher implements Pos
           to: anyPosition.dataProps.to,
           liquidity: totalLiquidity,
           positionId: anyPosition.dataProps.positionId,
+          assetStandard: Standard.ERC_721,
         },
         displayProps: {
           label: anyPosition.displayProps.label,