+++ src/apps/redacted-cartel/ethereum/redacted-cartel.bond.contract-position-fetcher.ts
@@ -40,9 +40,21 @@ export class EthereumRedactedCartelBondContractPositionFetcher extends ContractP
     const [principle, claimable] = await Promise.all([contract.principal(), contract.BTRFLY()]);
 
     return [
-      { address: claimable, metaType: MetaType.VESTING },
-      { address: claimable, metaType: MetaType.CLAIMABLE },
-      { address: principle, metaType: MetaType.SUPPLIED },
+      {
+        metaType: MetaType.VESTING,
+        address: claimable,
+        network: this.network,
+      },
+      {
+        metaType: MetaType.CLAIMABLE,
+        address: claimable,
+        network: this.network,
+      },
+      {
+        metaType: MetaType.SUPPLIED,
+        address: principle,
+        network: this.network,
+      },
     ];
   }