+++ src/apps/sushiswap-kashi/arbitrum/sushiswap-kashi.lending.token-fetcher.ts
@@ -1,29 +1,9 @@
-import { Inject } from '@nestjs/common';
+import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
 
-import { Register } from '~app-toolkit/decorators';
-import { PositionFetcher } from '~position/position-fetcher.interface';
-import { AppTokenPosition } from '~position/position.interface';
-import { Network } from '~types/network.interface';
+import { SushiswapKashiLendingTokenFetcher } from '../common/sushiswap-kashi.lending.token-fetcher';
 
-import { SushiswapKashiLendingTokenHelper } from '../helpers/sushiswap-kashi.lending.token-helper';
-import { SUSHISWAP_KASHI_DEFINITION } from '../sushiswap-kashi.definition';
-
-const appId = SUSHISWAP_KASHI_DEFINITION.id;
-const groupId = SUSHISWAP_KASHI_DEFINITION.groups.lending.id;
-const network = Network.ARBITRUM_MAINNET;
-
-@Register.TokenPositionFetcher({ appId, groupId, network })
-export class ArbitrumSushiswapKashiLendingTokenFetcher implements PositionFetcher<AppTokenPosition> {
-  constructor(
-    @Inject(SushiswapKashiLendingTokenHelper) private readonly lendingTokenHelper: SushiswapKashiLendingTokenHelper,
-  ) {}
-
-  getPositions() {
-    return this.lendingTokenHelper.getTokens({
-      network,
-      subgraphUrl: 'https://api.thegraph.com/subgraphs/name/sushiswap/arbitrum-bentobox',
-      subgraphVersion: 1,
-      first: 1000,
-    });
-  }
+@PositionTemplate()
+export class ArbitrumSushiswapKashiLendingTokenFetcher extends SushiswapKashiLendingTokenFetcher {
+  groupLabel = 'Lending';
+  subgraphUrl = 'https://api.thegraph.com/subgraphs/name/matthewlilley/kashi-arbitrum';
 }