+++ src/apps/unipilot/ethereum/unipilot.pool.token-fetcher.ts
@@ -1,0 +1,8 @@
+import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
+
+import { UnipilotVaultTokenFetcher } from '../common/unipilot.vault.token-fetcher';
+
+@PositionTemplate()
+export class EthereumUnipilotPoolTokenFetcher extends UnipilotVaultTokenFetcher {
+  groupLabel = 'Vaults';
+}